+++ shared_src/Board.h
@@ -154,7 +154,7 @@ es_status_codes GetDSC(uint32_t drvno, uint8_t DSCNumber, uint32_t* ADSC, uint32
 void PollDmaBufferToUserBuffer(uint32_t* drvno_p);
 void GetCurrentScanNumber(uint32_t drvno, int64_t* sample, int64_t* block);
 void GetScanNumber(uint32_t drvno, int64_t offset, int64_t* sample, int64_t* block);
-es_status_codes SetTicnt(uint32_t drvno, uint8_t divider);
+es_status_codes SetSticnt(uint32_t drvno, uint8_t divider);
 es_status_codes SetTocnt(uint32_t drvno, uint8_t divider);
 es_status_codes GetIsTdc(uint32_t drvno, bool* isTdc);
 es_status_codes GetIsDsc(uint32_t drvno, bool* isDsc);